--- release-notifier/internal/cache/cache.go
@@ -1,43 +1,0 @@
-/*******************************************************************************
- * Copyright (c) 2023 Contributors to the Eclipse Foundation
- *
- * See the NOTICE file(s) distributed with this work for additional
- * information regarding copyright ownership.
- *
- * This program and the accompanying materials are made available under the
- * terms of the Apache License, Version 2.0 which is available at
- * https://www.apache.org/licenses/LICENSE-2.0.
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
- * WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
- * License for the specific language governing permissions and limitations
- * under the License.
- *
- * SPDX-License-Identifier: Apache-2.0
- ******************************************************************************/
-package cache
-
-import (
-	"log"
-	"os"
-)
-
-func GetPrevPSQLRelFromArtifact(artifact string) string {
-	data, err := os.ReadFile(artifact)
-
-	if err != nil {
-		return ""
-	}
-
-	release := string(data)
-	return release
-}
-
-func SaveLatestPSQLRel(release string) {
-	err := os.WriteFile("psql_release", []byte(release), 0644)
-
-	if err != nil {
-		log.Fatalln(err)
-	}
-}
