+++ release-notifier/templates/mail-psql.html.tmpl
